requirements.txt
dash==2.13.0
dash-bootstrap-components==1.5.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
gunicorn==21.2.0
geopandas>=0.13.0
shapely>=2.0.1
statsmodels>=0.14.0
scikit-learn>=1.3.0
scipy>=1.11.2
python-dotenv>=1.0.0
werkzeug>=2.3.7
flask>=2.2.5